src/containers/autopilot/requirements.txt
nats-py~=2.2
pynmea2~=1.18
pymodbus~=3.0.2
